+++ twitter4j-v2-support/src/main/kotlin/twitter4j/JSONObjectEx.kt
@@ -7,3 +7,7 @@ internal fun JSONObject.optLongOrNull(name: String): Long? {
 internal fun JSONObject.optIntOrNull(name: String): Int? {
     return optInt(name, -1).takeIf { it != -1 }
 }
+
+internal fun JSONObject.optBooleanOrNull(name: String): Boolean? {
+    return if (!has(name)) null else optBoolean(name, false)
+}